setup/⚙️ adding new environment and finishing typescript config
