--- .babelrc
@@ -1,6 +1,0 @@
-{
-    "presets": [
-        "@babel/preset-env"
-    ],
-    "plugins": ["@babel/plugin-transform-runtime", "@babel/plugin-transform-regenerator"]
-}
