+++ .env.example
@@ -1,4 +1,14 @@
 MONGODB_URL=
 MONGODB_TEST_URL=
 ASTROLOGY_API_URL=
-JWT_KEY=J1N7@U2020
+JWT_KEY=
+GOOGLE_API_MAPS_KEY=
+
+FIREBASE_BUCKET_PROJECT_ID=
+FIREBASE_BUCKET_PROJECT_KEY_ID=
+FIREBASE_BUCKET_PRIVATE_KEY=
+FIREBASE_BUCKET_EMAIL=
+FIREBASE_BUCKET_CLIENT_ID=
+FIREBASE_BUCKET_X509_CERT_URL=
+
+BUCKET_ID=
