+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "yarn run jest",
-    "dev": "ts-node -r dotenv/config ./src/index.js ",
+    "dev": "node --abort-on-uncaught-exception -r dotenv/config -r tsconfig-paths/register -r ts-node/register ./src/index.js",
     "start": "env-cmd -f ./.env node ./lib",
     "build": "rm -rf lib && ./node_modules/.bin/babel src --out-dir lib"
   },
@@ -19,6 +19,7 @@
     "apollo-server-integration-testing": "^2.3.0",
     "apollo-server-testing": "^2.11.0",
     "axios": "^0.19.2",
+    "base-64": "^1.0.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
@@ -39,16 +40,10 @@
     "validator": "^12.2.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.10.1",
-    "@babel/core": "^7.10.2",
-    "@babel/node": "^7.10.1",
-    "@babel/plugin-transform-regenerator": "^7.10.1",
-    "@babel/plugin-transform-runtime": "^7.10.1",
-    "@babel/preset-env": "^7.10.2",
-    "@react-native-community/eslint-config": "^2.0.0",
+    "@types/base-64": "^1.0.0",
+    "@types/node": "^16.4.0",
     "@typescript-eslint/eslint-plugin": "^4.28.4",
     "@typescript-eslint/parser": "^4.28.4",
-    "babel-eslint": "^10.1.0",
     "eslint": "^7.8.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^3.4.0",
@@ -58,18 +53,11 @@
     "eslint-plugin-jest": "^21.18.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^2.6.2",
-    "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "eslint-plugin-react-native": "^3.2.1",
     "jest": "^25.1.0",
-    "nodemon": "^2.0.2",
     "prettier": "^1.19.1",
     "ts-node": "^10.4.0",
-    "ts-node-dev": "^1.1.8",
+    "tsconfig-paths": "^3.10.1",
     "typescript": "^4.3.5"
-  },
-  "resolutions": {
-    "@types/react": "^16"
   },
   "jest": {
     "testEnvironment": "node",
