--- src/storage/configs.js
@@ -1,14 +1,0 @@
-export const config = {
-  type: 'service_account',
-  project_id: 'jintou-d0ad5',
-  private_key_id: 'd3c43a5c474ab9f20599c3000c3eba1346c1d34a',
-  private_key:
-    '-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDDXVpVT5dnix6H\nsljSkaiGIxqDQ52mV0Z5AqCxDoPgLyVRFR+gaiMjQgKh1IvaCam1N+itIspbvVBq\nE98T6xsyBvhA7U0SIW98csTYk51O62XdT+MrJyMfvzYN+3yLNWHVonI/RQ4pJyxE\nIuP4WBj4XNR6Uhy6tA/SK5kejtyZ+04CRBbGAkTJgHhWa/lKEcXyHeiLzRZ5kqrN\nUwuLxTa5NmAD0Xw/Q8LxZSs6Z1v4IZxehiP9zvKAwMQXFZ2h4NTeC2RoHJXLKRfu\n8K7bwHFoPl/86GUeZJL1Lu6m3GNnjc7K2PRotuQr0rmbHZp9QCK4ZMqggp/qyd6h\nM9wXDMLpAgMBAAECggEAB7a6511s0VDOlCBi83ApjZ+l31hJyZSZRWk+Jl+NWMpj\nwyTTX9vGUWZngN5J5gJgVLXFa837hs3IdrfGz3GGsbNRarbem9WoIkDMgZn/1Grl\nAG0JajMcZuzU2ePRJ2n51/FBvjp1dFiF80Fxa4/cRIozIDVNHxjBpdRvh6OvwCtN\n4p/7uwU/e6bCSCygjrOg5reJHPU01qsbXUYBSa8gabUfWcomRXPMgpZK18BnivEm\nlvxRJfHlITk61zAWQyhz+wNed9YGXlaDTyiZrpcDn2BVdlnxjd8U+12YqT9FyjSU\n3TNT11dkmEyodnug+3DEHFyAF+WQY5jNPMXQbHriAQKBgQD/ph1YeMXCn+1iPgxg\nN2cziS+Xpa9cPgTnSQBgW3pEwyxuWjxSXa8V/KxFZUjLoxIWlUk4CNRpuZYhWBOH\nclD+uO6hHM53pA9XwPwMlnCG0lurXE2UKuBAp4D1Mwemi+QdbicIIrJFmLNi+dqy\npfIESghSN2DWo9MIZezUfhbnQQKBgQDDogrgKyxdqrinVQ9bzV0ENk0TWRuriYy0\nimuLMuPREQbng6SGWtzuiriO94XZ9+5cc9ZAUawsdrux6DyLEJd5M+rap48P514u\nBnVK7dJJQbMixL2Ebhhel7BHuOQq0mSAFUfuUV82oPsN5t3Z6vv8FulvTX+LXRKn\nHdtfnejZqQKBgA18NLYTANOIXhw9i6eTpJT6WpFUQxDJkLamiieLC8MOzIwxpFbv\n7Clrv6aVdiaPvoeEbm1PQe0lxf6SR81wQUd+d8/11gpHlCub2W/Mi7kA5ab/YnQ7\nYsLmUw4cF/44CNtokRSDyAWVzXieF2QvDrodMZ0soKYR1DLwNXLsB8dBAoGBAIgd\n1BFbGLhnU6lhvpjyT5jhq0keA+mbf+ISOXeVxfsuLK6i4rN0vyneLUGUKXaaUifT\nlSwVZ4Vca4YCfzWHKs1y+SJ/2UOIiPffGHc0985uvxQe7XoaHCycwFN26pJ85t5T\ncEP6LBRXd7aZfpzkKcWhr3P46Rhvfs32MS/bMCtZAoGBAMK3YWn14r23cUnGV3aA\n20745gC1WpqWYo2Vz8xQnX1mR2eXQ2Hoh7N9h0Rk2kOUDLMGzkWy+Luhvlli8PDk\njevB9DcPfoWSDg8pPxg6RgDZZbA93Lb6IN2OTMC5XTm4ku7oWYm9Gda8KYFTqGBk\ncmTQ1vA2JnLEQU/BeC9ZlJXn\n-----END PRIVATE KEY-----\n',
-  client_email: 'jintou-admin@jintou-d0ad5.iam.gserviceaccount.com',
-  client_id: '113278036013612771096',
-  auth_uri: 'https://accounts.google.com/o/oauth2/auth',
-  token_uri: 'https://oauth2.googleapis.com/token',
-  auth_provider_x509_cert_url: 'https://www.googleapis.com/oauth2/v1/certs',
-  client_x509_cert_url:
-    'https://www.googleapis.com/robot/v1/metadata/x509/jintou-admin%40jintou-d0ad5.iam.gserviceaccount.com'
-};
